rag_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sentence-transformers==2.3.1
faiss-cpu>=1.8.0
pandas==2.2.0
numpy==1.26.3
python-multipart
pyarrow
spacy>=3.7.0
